MB-5333 we don't know the type for the iov_len member
Reviewed-on: http://review.couchbase.org/16198
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>

+++ ep_testsuite.cc
@@ -352,7 +352,7 @@ static enum test_result check_key_value(ENGINE_HANDLE *h, ENGINE_HANDLE_V1 *h1,
     if (vlen != info.value[0].iov_len) {
         std::cerr << "Expected length " << vlen
                   << " got " << info.value[0].iov_len << std::endl;
-        checkeq(vlen, info.value[0].iov_len, "Length mismatch.");
+        checkeq(vlen, (size_t)info.value[0].iov_len, "Length mismatch.");
     }
 
     check(memcmp(info.value[0].iov_base, val, vlen) == 0,
